requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pathlib==1.0.1
pypdf==3.17.4
pillow==10.1.0
websockets==12.0
python-multipart==0.0.6 